Add alpha variable support
Well, this doesn't really work because there's no definition of
the value of the symbol referred to here.

This is modeled after the instructions in the tutorial as follows:

    MOVE A(PC),D0

I'm not sure what the 68k assembly does in this particular example,
and there isn't an example program and what the expected outpout is,
so it's unclear what the right thing to do here is.

So I'm committing this and hoping that it's correct. I'm sure I'll
stumble here in the next steps when we build on this, clearing up
what the expected behavior should be.

We'll find out.

If nothing else, it produces what I *hope* is the right Assembler:

    movl A(%rip), %eax