Define a data segment variable mid-text segment
This allows us to use _a(%rip) as expected since we now have an address
that that refers to.

Defaults to 0x0, but probably not what's desired: maybe an undefined
value that allows us to raise an error stating that the expected
action is to assign before use.

Also, what about typing? And how about not having to switch segments
mid-stream? (This would be addressed with buffering via AST or
something similar.)

Feels gross but it'll allow forward progress.