Add IO
The problem is this conflates our interpreter's IO with
the program's IO. We probably need to separate them.

So should we not be writing ASM anymore? If I remove all
that, this will be an interpreter, which I think was the
intention originally. May end up gutting the ASM bits and
renaming it, which would make it much more simple.

We'll see.